backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mysqlclient==2.2.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pytest==9.0.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
requests==2.32.5
scikit-learn==1.5.2
scipy==1.13.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.44
threadpoolctl==3.5.0
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
google-generativeai==0.8.5
